+++ naga/src/front/wgsl/parse/directive.rs
@@ -2,9 +2,16 @@
 //!
 //! See also <https://www.w3.org/TR/WGSL/#directives>.
 
+pub mod enable_extension;
+pub(crate) mod language_extension;
+
 /// A parsed sentinel word indicating the type of directive to be parsed next.
 #[derive(Clone, Copy, Debug, Hash, Eq, PartialEq)]
-pub enum DirectiveKind {
+pub(crate) enum DirectiveKind {
+    /// An [`enable_extension`].
+    Enable,
+    /// A [`language_extension`].
+    Requires,
     Unimplemented(UnimplementedDirectiveKind),
 }
 
@@ -17,8 +24,8 @@ impl DirectiveKind {
     pub fn from_ident(s: &str) -> Option<Self> {
         Some(match s {
             Self::DIAGNOSTIC => Self::Unimplemented(UnimplementedDirectiveKind::Diagnostic),
-            Self::ENABLE => Self::Unimplemented(UnimplementedDirectiveKind::Enable),
-            Self::REQUIRES => Self::Unimplemented(UnimplementedDirectiveKind::Requires),
+            Self::ENABLE => Self::Enable,
+            Self::REQUIRES => Self::Requires,
             _ => return None,
         })
     }
@@ -26,10 +33,10 @@ impl DirectiveKind {
     /// Maps this [`DirectiveKind`] into the sentinel word associated with it in WGSL.
     pub const fn to_ident(self) -> &'static str {
         match self {
+            Self::Enable => Self::ENABLE,
+            Self::Requires => Self::REQUIRES,
             Self::Unimplemented(kind) => match kind {
                 UnimplementedDirectiveKind::Diagnostic => Self::DIAGNOSTIC,
-                UnimplementedDirectiveKind::Enable => Self::ENABLE,
-                UnimplementedDirectiveKind::Requires => Self::REQUIRES,
             },
         }
     }
@@ -45,18 +52,14 @@ impl DirectiveKind {
 /// A [`DirectiveKind`] that is not yet implemented. See [`DirectiveKind::Unimplemented`].
 #[derive(Clone, Copy, Debug, Hash, Eq, PartialEq)]
 #[cfg_attr(test, derive(strum::EnumIter))]
-pub enum UnimplementedDirectiveKind {
+pub(crate) enum UnimplementedDirectiveKind {
     Diagnostic,
-    Enable,
-    Requires,
 }
 
 impl UnimplementedDirectiveKind {
     pub const fn tracking_issue_num(self) -> u16 {
         match self {
             Self::Diagnostic => 5320,
-            Self::Requires => 6350,
-            Self::Enable => 5476,
         }
     }
 }
@@ -78,39 +81,13 @@ mod test {
                 UnimplementedDirectiveKind::Diagnostic => {
                     shader = "diagnostic(off,derivative_uniformity);";
                     expected_msg = "\
-error: `diagnostic` is not yet implemented
+error: the `diagnostic` directive is not yet implemented
   ┌─ wgsl:1:1
   │
 1 │ diagnostic(off,derivative_uniformity);
   │ ^^^^^^^^^^ this global directive is standard, but not yet implemented
   │
   = note: Let Naga maintainers know that you ran into this at <https://github.com/gfx-rs/wgpu/issues/5320>, so they can prioritize it!
-
-";
-                }
-                UnimplementedDirectiveKind::Enable => {
-                    shader = "enable f16;";
-                    expected_msg = "\
-error: `enable` is not yet implemented
-  ┌─ wgsl:1:1
-  │
-1 │ enable f16;
-  │ ^^^^^^ this global directive is standard, but not yet implemented
-  │
-  = note: Let Naga maintainers know that you ran into this at <https://github.com/gfx-rs/wgpu/issues/5476>, so they can prioritize it!
-
-";
-                }
-                UnimplementedDirectiveKind::Requires => {
-                    shader = "requires readonly_and_readwrite_storage_textures";
-                    expected_msg = "\
-error: `requires` is not yet implemented
-  ┌─ wgsl:1:1
-  │
-1 │ requires readonly_and_readwrite_storage_textures
-  │ ^^^^^^^^ this global directive is standard, but not yet implemented
-  │
-  = note: Let Naga maintainers know that you ran into this at <https://github.com/gfx-rs/wgpu/issues/6350>, so they can prioritize it!
 
 ";
                 }
@@ -139,7 +116,7 @@ error: expected global declaration, but found a global directive
 
 ";
                 }
-                DirectiveKind::Unimplemented(UnimplementedDirectiveKind::Enable) => {
+                DirectiveKind::Enable => {
                     directive = "enable f16";
                     expected_msg = "\
 error: expected global declaration, but found a global directive
@@ -152,7 +129,7 @@ error: expected global declaration, but found a global directive
 
 ";
                 }
-                DirectiveKind::Unimplemented(UnimplementedDirectiveKind::Requires) => {
+                DirectiveKind::Requires => {
                     directive = "requires readonly_and_readwrite_storage_textures";
                     expected_msg = "\
 error: expected global declaration, but found a global directive